% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_core.R
\name{apply_move}
\alias{apply_move}
\title{Apply Move to Board}
\usage{
apply_move(board, move)
}
\arguments{
\item{board}{Current matrix}

\item{move}{A list structure (from get_legal_moves) containing from, to, and captures}
}
\value{
Updated board matrix
}
\description{
Executes a move, removes captured pieces, and handles promotion.
}
