% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_core.R
\name{check_game_state}
\alias{check_game_state}
\title{Check Game State}
\usage{
check_game_state(board, next_player)
}
\arguments{
\item{board}{Current matrix}

\item{next_player}{The player who is about to move (1 or 2)}
}
\value{
Character: "active", "white_won", "black_won"
}
\description{
Determines if the game is over based on the \emph{next} player's available moves.
If the next player has no moves, they lose.
}
