% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_ai.R
\name{get_ai_move}
\alias{get_ai_move}
\title{Get AI Move}
\usage{
get_ai_move(board, player, level = 3)
}
\arguments{
\item{board}{Matrix 8x8}

\item{player}{1 (White) or 2 (Black)}

\item{level}{AI difficulty: 1 (Beginner), 2 (Amateur), 3 (Advanced)}
}
\value{
Selected move object
}
\description{
Main entry point for AI move selection
}
