% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_moves.R
\name{get_legal_moves}
\alias{get_legal_moves}
\title{Get All Legal Moves}
\usage{
get_legal_moves(board, player)
}
\arguments{
\item{board}{Matrix 8x8}

\item{player}{1 (White) or 2 (Black)}
}
\value{
List of moves.
}
\description{
Entry point for move generation. Enforces mandatory capture.
}
