% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\docType{data}
\name{i18n_dict}
\alias{i18n_dict}
\title{Dictionary for rcheckers}
\format{
An object of class \code{list} of length 41.
}
\usage{
i18n_dict
}
\description{
Dictionary for rcheckers
}
\keyword{internal}
