% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_board.R, R/logic_core.R
\name{init_board}
\alias{init_board}
\title{Initialize Game Board}
\usage{
init_board()

init_board()
}
\value{
A matrix 8x8.

A matrix 8x8.
}
\description{
Creates a standard 8x8 Russian Checkers board.
1=White, 2=Black, 3=White King, 4=Black King.

Creates a standard 8x8 Russian Checkers board.
0=Empty, 1=White Man, 2=Black Man, 3=White King, 4=Black King.
Top of matrix (row 1) is Black side, Bottom (row 8) is White side.
}
