% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_ai.R
\name{get_ai_move}
\alias{get_ai_move}
\title{Get Best Move for AI}
\usage{
get_ai_move(board, player)
}
\arguments{
\item{board}{Current board}

\item{player}{AI player color (usually 2)}
}
\value{
A move object (list) or NULL if no moves
}
\description{
Currently implements a simple strategy:
\enumerate{
\item Must capture if available.
\item If multiple captures, picks the one removing most pieces.
\item If quiet move, picks randomly (can be improved later).
}
}
